+++ terraform/modules/network/locals.tf
@@ -1,0 +1,15 @@
+locals {
+
+  name_prefix = "${var.project_name}-${var.environment}"
+
+  common_tags = merge(
+    {
+      Project     = var.project_name
+      Environment = var.environment
+      ManagedBy   = "Terraform"
+      Module      = "Network"
+    },
+    var.tags
+  )
+
+}