+++ terraform/modules/network/subnets.tf
@@ -1,0 +1,58 @@
+resource "aws_subnet" "public" {
+
+  count = length(var.public_subnet_cidrs)
+
+  vpc_id = aws_vpc.this.id
+
+  cidr_block = var.public_subnet_cidrs[count.index]
+
+  availability_zone = var.availability_zones[count.index]
+
+  map_public_ip_on_launch = true
+
+  tags = merge(
+    local.common_tags,
+    {
+      Name = "${local.name_prefix}-public-${count.index + 1}"
+      Tier = "Public"
+    }
+  )
+}
+
+resource "aws_subnet" "private" {
+
+  count = length(var.private_subnet_cidrs)
+
+  vpc_id = aws_vpc.this.id
+
+  cidr_block = var.private_subnet_cidrs[count.index]
+
+  availability_zone = var.availability_zones[count.index]
+
+  tags = merge(
+    local.common_tags,
+    {
+      Name = "${local.name_prefix}-private-${count.index + 1}"
+      Tier = "Private"
+    }
+  )
+}
+
+resource "aws_subnet" "database" {
+
+  count = length(var.database_subnet_cidrs)
+
+  vpc_id = aws_vpc.this.id
+
+  cidr_block = var.database_subnet_cidrs[count.index]
+
+  availability_zone = var.availability_zones[count.index]
+
+  tags = merge(
+    local.common_tags,
+    {
+      Name = "${local.name_prefix}-database-${count.index + 1}"
+      Tier = "Database"
+    }
+  )
+}